functions/fn1/requirements.txt
firebase_functions~=0.4.2
gspread~=6.1.4
pandas~=2.2.3
google-cloud-firestore~=2.19.0
google-auth~=2.36.0
requests~=2.32.3
pillow~=11.0.0